requirements.txt
mcp>=0.9.0
httpx>=0.27.0
python-dotenv>=1.0.0
flask>=3.0.0
anthropic>=0.39.0

# Data Science & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Deep Learning
torch>=2.0.0
tensorflow>=2.13.0

# NLP & Sentiment Analysis
transformers>=4.30.0
sentencepiece>=0.1.99
nltk>=3.8.0
textblob>=0.17.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Backtesting
backtrader>=1.9.78.123

# Technical Analysis
ta-lib>=0.4.28
pandas-ta>=0.3.14b

# Utilities
joblib>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0
